// src/lib/pb.d.ts
export enum Collections {
	ActiveMessage = "activeMessage",
	Categories = "categories",
	DisplayMessages = "displayMessages",
	Drinks = "drinks",
	OrderDrink = "order_drink",
	Orders = "orders",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	created: IsoDateString
	updated: IsoDateString
	collectionId: string
	collectionName: Collections
	expand?: T
}

export type AuthSystemFields<T = never> = {
	email: string
	emailVisibility: boolean
	username: string
	verified: boolean
} & BaseSystemFields<T>

// Record types for each collection

export type ActiveMessageRecord = {
	isVisible?: boolean
	message?: RecordIdString
}

export type CategoriesRecord = {
	name: string
	sort_order: number
}

export type DisplayMessagesRecord = {
	subtext?: string
	title?: string
}

export type DrinksRecord = {
	category: RecordIdString
	image: string
	name: string
	price: number
}
// ...
export type CategoriesResponse<Texpand = unknown> = Required<CategoriesRecord> & BaseSystemFields<Texpand>
export type DisplayMessagesResponse<Texpand = unknown> = Required<DisplayMessagesRecord> & BaseSystemFields<Texpand>
export type DrinksResponse<Texpand = unknown> = Required<DrinksRecord> & BaseSystemFields<Texpand>